+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/container/ContainerNetwork.java
@@ -135,7 +135,9 @@ public class ContainerNetwork extends AbstractNetwork {
                     .filter(receiver -> !receiver.equals(sender))
                     .map(receiver -> networkBehavior.getProxyEndpoint(sender, receiver))
                     .toList();
-            sender.configuration().setNetworkEndpoints(GossipConfig_.ENDPOINT_OVERRIDES, endpointOverrides);
+            ((ContainerNode) sender)
+                    .configuration()
+                    .setNetworkEndpoints(GossipConfig_.ENDPOINT_OVERRIDES, endpointOverrides);
         }
     }
