+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/AbstractNodeConfiguration.java
@@ -4,14 +4,9 @@ package org.hiero.otter.fixtures.internal;
 import static java.util.Objects.requireNonNull;
 import static org.hiero.otter.fixtures.internal.helpers.Utils.createConfiguration;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator.Feature;
 import com.swirlds.component.framework.schedulers.builders.TaskSchedulerConfiguration;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.platform.config.PathsConfig_;
-import com.swirlds.platform.gossip.config.NetworkEndpoint;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.nio.file.Path;
 import java.time.Duration;
@@ -19,18 +14,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 import org.hiero.otter.fixtures.NodeConfiguration;
 import org.hiero.otter.fixtures.internal.AbstractNode.LifeCycle;
-import org.jetbrains.annotations.NotNull;
 
 /**
  * An abstract base class for node configurations that provides common functionality
  */
 public abstract class AbstractNodeConfiguration implements NodeConfiguration {
-
-    private static final ObjectMapper OBJECT_MAPPER =
-            new ObjectMapper(new YAMLFactory().disable(Feature.WRITE_DOC_START_MARKER));
 
     protected final Map<String, String> overriddenProperties = new HashMap<>();
 
@@ -130,7 +120,7 @@ public abstract class AbstractNodeConfiguration implements NodeConfiguration {
      */
     @Override
     @NonNull
-    public NodeConfiguration setStrings(@NonNull final String key, @NonNull final List<String> values) {
+    public NodeConfiguration set(@NonNull final String key, @NonNull final List<String> values) {
         throwIfNodeIsRunning();
         overriddenProperties.put(key, String.join(",", values));
         return this;
@@ -144,21 +134,6 @@ public abstract class AbstractNodeConfiguration implements NodeConfiguration {
     public NodeConfiguration set(@NonNull final String key, @NonNull final Path path) {
         throwIfNodeIsRunning();
         overriddenProperties.put(key, path.toString());
-        return this;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @NonNull
-    public NodeConfiguration setNetworkEndpoints(
-            @NotNull final String key, @NotNull final List<NetworkEndpoint> endpoints) {
-        throwIfNodeIsRunning();
-        final String value = endpoints.stream()
-                .map(AbstractNodeConfiguration::convertEndpoint)
-                .collect(Collectors.joining(","));
-        overriddenProperties.put(key, value);
         return this;
     }
 
@@ -196,16 +171,7 @@ public abstract class AbstractNodeConfiguration implements NodeConfiguration {
         return this;
     }
 
-    private static String convertEndpoint(@NonNull final NetworkEndpoint endpoint) {
-        try {
-            return OBJECT_MAPPER.writeValueAsString(endpoint).replaceAll("\"", "\\\"");
-        } catch (final JsonProcessingException e) {
-            // This should not happen as the list is expected to be serializable
-            throw new RuntimeException("Exception while serializing endpoints", e);
-        }
-    }
-
-    private void throwIfNodeIsRunning() {
+    protected final void throwIfNodeIsRunning() {
         if (lifecycleSupplier.get() == LifeCycle.RUNNING) {
             throw new IllegalStateException("Configuration modification is not allowed when the node is running.");
         }
